requirements.txt
albumentations==1.3.0
Flask==2.2.2
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
imutils==0.5.4
keras==2.9.0
matplotlib==3.5.3
numpy==1.23.2
onnx==1.13.1
onnx_graphsurgeon==0.0.1.dev5
pafy==0.5.5
pandas==1.4.4
Pillow==9.4.0
prediction_utils==0.0.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.1
scipy==1.9.1
seaborn==0.12.2
setuptools==65.6.3
SQLAlchemy_Utils==0.38.3
tensorflow==2.11.0
tensorflow_gpu==2.9.1
thop==0.1.1.post2209072238
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
validators==0.20.0
wandb==0.14.0
Werkzeug==2.2.2
~pencv_python==4.6.0.66
